test: extract StartupPerformanceTracker reflection ctor to TestTrackerFactory
